Move auth to contrib.auth
