+++ starlette_web/tests/core/test_auth_backends.py
@@ -6,10 +6,10 @@ from starlette_web.common.http.exceptions import (
     AuthenticationFailedError,
     PermissionDeniedError,
 )
-from starlette_web.auth.backend import JWTAuthenticationBackend
-from starlette_web.auth.models import User, UserSession
-from starlette_web.auth.permissions import IsSuperuserPermission
-from starlette_web.auth.utils import encode_jwt, TOKEN_TYPE_RESET_PASSWORD, TOKEN_TYPE_REFRESH
+from starlette_web.contrib.auth.backend import JWTAuthenticationBackend
+from starlette_web.contrib.auth.models import User, UserSession
+from starlette_web.contrib.auth.permissions import IsSuperuserPermission
+from starlette_web.contrib.auth.utils import encode_jwt, TOKEN_TYPE_RESET_PASSWORD, TOKEN_TYPE_REFRESH
 from starlette_web.common.http.requests import PRequest
 from starlette_web.tests.helpers import await_
